Move crc32 C fallbacks to arch/generic
