+++ arch/generic/crc32_fold_c.c
@@ -4,10 +4,9 @@
  */
 #include "zbuild.h"
 #include "functable.h"
+#include "crc32.h"
 
-#include "crc32_fold.h"
-
-#include <limits.h>
+#include "crc32_fold_c.h"
 
 Z_INTERNAL uint32_t crc32_fold_reset_c(crc32_fold *crc) {
     crc->value = CRC32_INITIAL_VALUE;
