+++ deflate.h
@@ -12,7 +12,7 @@
 
 #include "zutil.h"
 #include "zendian.h"
-#include "crc32_fold.h"
+#include "crc32.h"
 
 /* define NO_GZIP when compiling if you want to disable gzip header and
    trailer creation by deflate().  NO_GZIP would be used to avoid linking in
