+++ functable.h
@@ -7,7 +7,7 @@
 #define FUNCTABLE_H_
 
 #include "deflate.h"
-#include "crc32_fold.h"
+#include "crc32_fold_c.h"
 #include "adler32_fold_c.h"
 
 #ifdef ZLIB_COMPAT
